% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afdc_add_connector_type_indicators.R
\name{afdc_add_connector_type_indicators}
\alias{afdc_add_connector_type_indicators}
\title{Add indicators for charger types to dataframe of ev stations}
\usage{
afdc_add_connector_type_indicators(ev_df)
}
\arguments{
\item{ev_df}{Dataframe of ev stations with "ev_connector_types" column}
}
\value{
ev_df Dataframe of ev stations with added columns indicating presence of charger connection types
}
\description{
Add indicators for charger types to dataframe of ev stations
}
