% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afdc_construct_query_url.R
\name{afdc_construct_query_url}
\alias{afdc_construct_query_url}
\title{Construct URL for AFDC API data request}
\usage{
afdc_construct_query_url(fuel_type = "ELEC", wh_state = "CO", limit = "all")
}
\arguments{
\item{fuel_type}{Fuel type, default = "ELEC"}

\item{wh_state}{State to get data for}

\item{limit}{Number of rows to return; default is "all"}
}
\value{
query_url URL for API request
}
\description{
Construct URL for AFDC API data request
}
