% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afdc_ev_get_stations.R
\name{afdc_ev_get_stations}
\alias{afdc_ev_get_stations}
\title{Fetch processed EV charging station data from NREL API}
\usage{
afdc_ev_get_stations(fuel_type = "ELEC", wh_state = "CO", limit = "all")
}
\arguments{
\item{fuel_type}{Fuel type, default = "ELEC"}

\item{wh_state}{State to get data for}

\item{limit}{Number of rows to return; default is "all"}
}
\value{
df Dataframe with EV charging station data
}
\description{
Fetch processed EV charging station data from NREL API
}
