% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afdc_get_api_request.R
\name{afdc_get_api_request}
\alias{afdc_get_api_request}
\title{Make API GET request and return data}
\usage{
afdc_get_api_request(query_url)
}
\arguments{
\item{query_url}{URL for the API query}
}
\value{
df: data from the API response
}
\description{
Make API GET request and return data
}
