% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afdc_process_ev_stations.R
\name{afdc_process_ev_stations}
\alias{afdc_process_ev_stations}
\title{Process AFDC station data for ev charging only}
\usage{
afdc_process_ev_stations(ev_df)
}
\arguments{
\item{ev_df}{Dataframe of EV charging station data as returned from AFDC API}
}
\value{
ev_df Dataframe with only EV-related fields
}
\description{
Process AFDC station data for ev charging only
}
